+++ scripts/tools/note.js
@@ -9,12 +9,29 @@ function getNoteCount(notes) {
     let nr = 0;
     let nb = 0;
     let b = 0;
+    let cn = 0;
+    let cb = 0;
     for (let i = notes.length - 1; i >= 0; i--) {
-        if (notes[i]._type === 0) nr++;
-        else if (notes[i]._type === 1) nb++;
-        else b++;
+        if (notes[i]._type === 0) {
+            nr++;
+            if (notes[i]._type._customData?._color) {
+                cn++;
+            }
+        }
+        else if (notes[i]._type === 1) {
+            nb++;
+            if (notes[i]._type._customData?._color) {
+                cn++;
+            }
+        }
+        else if (notes[i]._type === 3) {
+            b++;
+            if (notes[i]._type._customData?._color) {
+                cb++;
+            }
+        }
     }
-    return {red: nr, blue: nb, bomb: b};
+    return {red: nr, blue: nb, bomb: b, chromaN: cn, chromaB: cb};
 }
 function countNoteLayer(notes, l) {
     let count = 0;
@@ -372,6 +389,17 @@ function checkEndNote(n1, n2) {
     return false;
 }
 
+function checkForDouble(n1, notes, index) {
+    for (let i = index, len = notes.length; i < len; i++) {
+        if (notes[i]._time < n1._time + 0.01 && notes[i]._type != n1._type) {
+            return true;
+        }
+        else if (notes[i]._time > n1._time + 0.01) {
+            return false;
+        }
+    }
+}
+
 // check if note occupies post-swing space
 // also fuck dot note
 // i need to rewrite this
@@ -391,7 +419,7 @@ function detectHitboxStaircase(diff, mapSettings) {
             if (lastRed) { // nested if moment
                 if (swingNext(note, lastRed)) {
                     if (lastBlue && note._time - lastBlue._time !== 0 && isBelowTH(note._time - lastBlue._time, tool.hitbox.staircase)) {
-                        if (note._lineIndex === blueIndexOccupy && note._lineLayer === blueLayerOccupy) {
+                        if (note._lineIndex === blueIndexOccupy && note._lineLayer === blueLayerOccupy && !checkForDouble(note, notes, i)) {
                             arr.push(adjustTime(note._time, bpm, offset, bpmc));
                         }
                     }
@@ -428,7 +456,7 @@ function detectHitboxStaircase(diff, mapSettings) {
             if (lastBlue) {
                 if (swingNext(note, lastBlue)) {
                     if (lastRed && note._time - lastRed._time !== 0 && isBelowTH(note._time - lastRed._time, tool.hitbox.staircase)) {
-                        if (note._lineIndex === redIndexOccupy && note._lineLayer === redLayerOccupy) {
+                        if (note._lineIndex === redIndexOccupy && note._lineLayer === redLayerOccupy && !checkForDouble(note, notes, i)) {
                             arr.push(adjustTime(note._time, bpm, offset, bpmc));
                         }
                     }